This commit brings two new features to Spigot:
- Handle arrays of data
Now if you pass an array of values, Spigot will iterate over the
array of values and parse each one individually

- Associations
You can now specify a class as a mapping value in Spigot.define.
If you provide a class which also has a spigot resource map, Spigot
will use the map for the specified resource to parse nested data.